web/pyproject.toml
[project]
name = "neurotechx-website"
version = "0.1.0"
description = "Website and API services for NeurotechX"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.10",
    "whitenoise>=6.11.0",
]

[tool.ruff.format]
# Enable Black-style formatting
quote-style = "double"
indent-style = "space"
line-ending = "auto"
docstring-code-format = true

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "ruff>=0.13.1",
]
